projects/001-timestamp-microservice/solution/enricozoboli/requirements.txt
astroid==2.9.3
autopep8==1.6.0
backports.zoneinfo==0.2.1
dateparser==1.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.1
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.1.18
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
wrapt==1.13.3
